Add ability to locally publish mkdocs website (#394)

+++ deploy_website.sh
@@ -6,22 +6,25 @@
 # Install the packages with the following command:
 # pip install mkdocs mkdocs-material
 
-set -ex
+if [ "$1" = "--local" ]; then local=true; fi
+if ! [ $local ]; then
+  set -ex
 
-REPO="git@github.com:uber/AutoDispose.git"
-DIR=temp-clone
+  REPO="git@github.com:uber/AutoDispose.git"
+  DIR=temp-clone
 
-# Delete any existing temporary website clone
-rm -rf $DIR
+  # Delete any existing temporary website clone
+  rm -rf $DIR
 
-# Clone the current repo into temp folder
-git clone $REPO $DIR
+  # Clone the current repo into temp folder
+  git clone $REPO $DIR
 
-# Move working directory into temp folder
-cd $DIR
+  # Move working directory into temp folder
+  cd $DIR
 
-# Generate the API docs
-./gradlew dokka
+  # Generate the API docs
+  ./gradlew dokka
+fi
 
 # Copy in special files that GitHub wants in the project root.
 cat README.md > docs/index.md
@@ -30,8 +33,14 @@ cp CONTRIBUTING.md docs/contributing.md
 cp CODE_OF_CONDUCT.md docs/code-of-conduct.md
 
 # Build the site and push the new files up to GitHub
-mkdocs gh-deploy
+if ! [ $local ]; then
+  mkdocs gh-deploy
+else
+  mkdocs serve
+fi
 
 # Delete our temp folder
-cd ..
-rm -rf $DIR
+if ! [ $local ]; then
+  cd ..
+  rm -rf $DIR
+fi
